use wrapped grpc error
